+++ application/source/source/application_layer/service/handler/_in_contex_for/presentation_layer/service/actix_web_component/request_handler/application_programming_interface/version_1/mobile/_in_context_for/domain_layer/entity/entity/channel/_new_for_context/base/_new_for_contex/get_many_by_name/base/mod.rs
@@ -49,7 +49,7 @@ impl Base {
         }
 
         let channel_registry: Option<Vec<Channel>> = DataProviderChannelPostgresql::get_many_by_name(
-            &mut *ConnectionExtractor::get_postgresql_connection(&aggregate_connection_pool)?, channel_name.as_str(), &requery_channel_name, limit
+            &mut *ConnectionExtractor::get_postgresql_connection(&aggregate_connection_pool)?, channel_name.as_str(), &requery_channel_name, limit as i16
         )?;
 
         return Ok(Response::new(channel_registry));
